Remove `palette.text.highlight` and update components to use their corresponding colors
