+++ components/SwapProgress.tsx
@@ -100,7 +100,7 @@ const styles = {
 		line-height: 1;
 		text-align: center;
 		text-transform: uppercase;
-		color: ${palette.text.highlight};
+		color: ${palette.primary.main};
 	`,
 
 	message: css`
