+++ pages/pool.tsx
@@ -1,11 +1,11 @@
 import { Api } from "@cennznet/api";
 import { CENNZAsset } from "@/types";
 import fetchPoolAssets from "@/utils/fetchPoolAssets";
-import { css } from "@emotion/react";
-import { Theme } from "@mui/material";
 import { API_URL } from "@/constants";
 import generateGlobalProps from "@/utils/generateGlobalProps";
 import MainPanel from "@/components/MainPanel";
+import PoolProvider from "@/providers/PoolProvider";
+import { VFC } from "react";
 
 export async function getStaticProps() {
 	const api = await Api.create({ provider: API_URL });
@@ -18,10 +18,14 @@ export async function getStaticProps() {
 	};
 }
 
-const Pool: React.FC<{ supportedAssets: CENNZAsset[] }> = ({
-	supportedAssets,
-}) => {
-	return <MainPanel defaultTitle="Add to Pool"></MainPanel>;
+const Pool: VFC<{ supportedAssets: CENNZAsset[] }> = ({ supportedAssets }) => {
+	console.log({ supportedAssets });
+
+	return (
+		<PoolProvider supportedAssets={supportedAssets}>
+			<MainPanel defaultTitle="Add to Pool"></MainPanel>
+		</PoolProvider>
+	);
 };
 
 export default Pool;
