+++ pages/swap.tsx
@@ -1,8 +1,6 @@
 import { Api } from "@cennznet/api";
 import { CENNZAsset } from "@/types";
 import fetchSwapAssets from "@/utils/fetchSwapAssets";
-import { css } from "@emotion/react";
-import { Theme } from "@mui/material";
 import { API_URL } from "@/constants";
 import generateGlobalProps from "@/utils/generateGlobalProps";
 import SwapProvider from "@/providers/SwapProvider";
@@ -12,6 +10,7 @@ import SwapStats from "@/components/SwapStats";
 import SwapSettings from "@/components/SwapSettings";
 import SwapProgress from "@/components/SwapProgress";
 import MainPanel from "@/components/MainPanel";
+import { VFC } from "react";
 
 export async function getStaticProps() {
 	const api = await Api.create({ provider: API_URL });
@@ -24,9 +23,7 @@ export async function getStaticProps() {
 	};
 }
 
-const Swap: React.FC<{ supportedAssets: CENNZAsset[] }> = ({
-	supportedAssets,
-}) => {
+const Swap: VFC<{ supportedAssets: CENNZAsset[] }> = ({ supportedAssets }) => {
 	return (
 		<SwapProvider supportedAssets={supportedAssets}>
 			<MainPanel defaultTitle="Swap">
